limit modification of posts and comments to users

+++ src/services/comment.js
@@ -4,8 +4,9 @@ const CommentController = require('../controllers/comment');
 class CommentService extends Service {
   constructor(...args) {
     super(...args);
-    this.controller = new CommentController();
+    this.controller = new CommentController(this.user);
   }
+
   /**
    *
    * @param {string} id - the id of a single post
@@ -24,12 +25,25 @@ class CommentService extends Service {
   }
 
   async update(id, data) {
-    const { body }= data;
-    return this.controller.update(body,id);
+    const { userId } = await this.controller.getUserId(id);
+    if (this.user.id === userId) {
+      const { body } = data;
+      return this.controller.update(body, id);
+    }
+    else {
+      return this.response.status(401).send({ message: 'unauthorized' });
+    }
   }
 
   async delete(id) {
-    return this.controller.delete(id);
+    const { userId } = await this.controller.getUserId(id);
+    if (this.user.id === userId) {
+      return this.controller.delete(id);
+    }
+    else {
+      return this.response.status(401).send({ message: 'unauthorized' });
+    }
   }
 }
-  module.exports = CommentService;
+
+module.exports = CommentService;
